typescript remove warning
