+++ package.json
@@ -10,15 +10,20 @@
     "@material-ui/core": "4.12.3",
     "@material-ui/icons": "4.11.2",
     "@mui/material": "^5.9.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router-dom": "5.3.0",
-    "react-scripts": "4.0.0",
-    "sass": "^1.54.0"
+    "@types/jest": "^28.1.6",
+    "@types/node": "^18.6.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "^5.3.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.54.1"
   },
   "devDependencies": {
     "@babel/runtime": "7.13.8",
-    "typescript": "4.1.3"
+    "typescript": "^4.7.4"
   },
   "scripts": {
     "start": "react-scripts start",
