+++ src/index.tsx
@@ -1,0 +1,8 @@
+import { createRoot } from "react-dom/client";
+
+import App from "./App";
+import "./styles.scss";
+
+const container = document.getElementById("root");
+const root = createRoot(container!); // createRoot(container!) if you use TypeScript
+root.render(<App />);
