+++ src/page/about.scss
@@ -1,152 +1,151 @@
 .claims {
-    font-family: "Avenir Next Condensed" !important;
-    background-color: #0b1127;
-    // background-image: url("../img/background.png");
-    // background-repeat: repeat-x;
+  font-family: "Avenir Next Condensed" !important;
+  background-color: #0b1127;
+  // background-image: url("../img/background.png");
+  // background-repeat: repeat-x;
 }
 
 .claim {
-    padding-top: 20px;
-    padding-bottom: 20px;
-    margin: 0 auto !important;
+  padding-top: 20px;
+  padding-bottom: 20px;
+  margin: 0 auto !important;
 }
 
 .claim-gifts-title-name {
-    font-size: 20px;
-    padding: 30px 0px 10px;
-    color: #d0d1d5;
-    letter-spacing: 0.1em;
+  font-size: 20px;
+  padding: 30px 0px 10px;
+  color: #d0d1d5;
+  letter-spacing: 0.1em;
 }
 
 .claim-content {
-    margin-bottom: 40px;
+  margin-bottom: 40px;
 }
 
 .connects-wallet {
-    display: flex;
-    justify-content: center;
+  display: flex;
+  justify-content: center;
 }
 
 .connects-button {
-    background-color: #2e3346 !important;
-    // color: #d0d1d5 !important;
-    color: #ffffff !important;
-    font-size: 20px !important;
+  background-color: #2e3346 !important;
+  // color: #d0d1d5 !important;
+  color: #ffffff !important;
+  font-size: 20px !important;
 }
 
 .total-balance {
-    padding-top: 10px;
+  padding-top: 10px;
 }
 
 .balance-title {
-    display: flex;
-    justify-content: center;
-    // color: #383d4e !important;
-    color: #838487 !important;
-    font-size: 25px;
-    margin: 0px;
-    letter-spacing: 0.3em;
-
+  display: flex;
+  justify-content: center;
+  // color: #383d4e !important;
+  color: #838487 !important;
+  font-size: 25px;
+  margin: 0px;
+  letter-spacing: 0.3em;
 }
 
 .balance-number {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 50px;
-    // color: #d0d1d5;
-    color: #ffffff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 50px;
+  // color: #d0d1d5;
+  color: #ffffff;
 }
 
 .balance-number-money {
-    padding-top: 2px;
+  padding-top: 2px;
 }
 
 .gifts-title {
-    display: flex;
-    justify-content: center;
-    // color: #383d4e;
-    color: #838487;
-    font-size: 24px;
-    letter-spacing: 0.1em;
+  display: flex;
+  justify-content: center;
+  // color: #383d4e;
+  color: #838487;
+  font-size: 24px;
+  letter-spacing: 0.1em;
 }
 
 .gifts-number {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: #d0d1d5;
-    font-size: 33px;
-    letter-spacing: 0.1em;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: #d0d1d5;
+  font-size: 33px;
+  letter-spacing: 0.1em;
 }
 
 p {
-    margin: 0px;
-    font-family: "Avenir Next Condensed" !important;
+  margin: 0px;
+  font-family: "Avenir Next Condensed" !important;
 }
 
 .MuiTypography-body1 {
-    line-height: 1.2 !important;
+  line-height: 1.2 !important;
 }
 
 .balance-number-img {
-    width: 20px;
-    height: 36px;
-    margin-left: 5px;
+  width: 20px;
+  height: 36px;
+  margin-left: 5px;
 }
 
 .gifts-number-img {
-    width: 15px;
-    height: 26px;
-    margin-left: 3px;
+  width: 15px;
+  height: 26px;
+  margin-left: 3px;
 }
 
 @media only screen and (max-width: 610px) {
-    .connects-button {
-        width: 50px;
-        font-size: 11px !important;
-    }
+  .connects-button {
+    width: 50px;
+    font-size: 11px !important;
+  }
 
-    .balance-title {
-        font-size: 18px;
-        letter-spacing: 0.1em;
-    }
+  .balance-title {
+    font-size: 18px;
+    letter-spacing: 0.1em;
+  }
 
-    .balance-number {
-        font-size: 37px;
-    }
+  .balance-number {
+    font-size: 37px;
+  }
 
-    .balance-number-money {
-        padding-top: 6px;
-    }
+  .balance-number-money {
+    padding-top: 6px;
+  }
 
-    .balance-number-img {
-        width: 15px;
-        height: 26px;
-        margin-left: 2px;
-    }
+  .balance-number-img {
+    width: 15px;
+    height: 26px;
+    margin-left: 2px;
+  }
 
-    .gifts-title {
-        font-size: 17px;
-        letter-spacing: 0.1em;
-    }
+  .gifts-title {
+    font-size: 17px;
+    letter-spacing: 0.1em;
+  }
 
-    .gifts-number {
-        font-size: 27px;
-        letter-spacing: 0.1em;
-    }
+  .gifts-number {
+    font-size: 27px;
+    letter-spacing: 0.1em;
+  }
 
-    .gifts-number-money {
-        padding-top: 4px;
-    }
+  .gifts-number-money {
+    padding-top: 4px;
+  }
 
-    .gifts-number-img {
-        width: 11px;
-        height: 19px;
-        margin-left: 1px;
-    }
+  .gifts-number-img {
+    width: 11px;
+    height: 19px;
+    margin-left: 1px;
+  }
 
-    .claim-gifts-title-name {
-        font-size: 15px;
-    }
+  .claim-gifts-title-name {
+    font-size: 15px;
+  }
 }
