--- src/page/contact.js
@@ -1,11 +1,0 @@
-import { Typography } from "@material-ui/core";
-
-import './contact.scss'
-
-const Contact = (props) => {
-  const { cointitle, coinmoney, gifttitle, giftmoney, imgurl } = props
-  return <Typography>
-
-  </Typography>;
-}
-export default Contact;
