+++ src/page/contact.tsx
@@ -1,0 +1,6 @@
+import { Typography } from "@material-ui/core";
+
+const Contact = () => {
+  return <Typography />;
+};
+export default Contact;
