+++ src/page/faq.tsx
@@ -1,0 +1,6 @@
+import { Typography } from "@material-ui/core";
+
+function Faq() {
+  return <Typography>This is the FAQ Page</Typography>;
+}
+export default Faq;
