+++ src/page/home.scss
@@ -1,52 +1,52 @@
 .home-total {
-    // background-color: #0b1127;
-    background-image: url("../img/background.png");
-    background-repeat: repeat-x;
-    height: 87.5vh;
-    display: flex;
-    padding-top: 30px;
+  // background-color: #0b1127;
+  background-image: url("../img/background.png");
+  background-repeat: repeat-x;
+  height: 87.5vh;
+  display: flex;
+  padding-top: 30px;
 }
 
 .home-title {
-    font-family: "Avenir Next Condensed" !important;
-    display: flex;
-    justify-content: center;
-    font-size: 80px;
-    letter-spacing: 0.3em;
-    color: #ffcc40;
-    margin: 0;
+  font-family: "Avenir Next Condensed" !important;
+  display: flex;
+  justify-content: center;
+  font-size: 80px;
+  letter-spacing: 0.3em;
+  color: #ffcc40;
+  margin: 0;
 }
 
 .home-logos {
-    display: flex;
-    justify-content: center;
-    margin: 52px 0px;
+  display: flex;
+  justify-content: center;
+  margin: 52px 0px;
 }
 
 .home-logo {
-    width: 420px;
-    height: 420px;
+  width: 420px;
+  height: 420px;
 }
 
 @media only screen and (max-width: 610px) {
-    .home-title {
-        font-size: 40px;
-    }
+  .home-title {
+    font-size: 40px;
+  }
 
-    .home-logos {
-        margin: 60px 0px;
-    }
+  .home-logos {
+    margin: 60px 0px;
+  }
 
-    .home-logo {
-        width: 212px;
-        height: 212px;
-    }
+  .home-logo {
+    width: 212px;
+    height: 212px;
+  }
 
-    // .home-total {
-    //     padding: 69px 0px;
-    // }
-    .home-total {
-        height: 87.5vh;
-        padding-top: 90px;
-    }
+  // .home-total {
+  //     padding: 69px 0px;
+  // }
+  .home-total {
+    height: 87.5vh;
+    padding-top: 90px;
+  }
 }
