+++ src/styles.scss
@@ -1,8 +1,3 @@
-@font-face {
-  font-family: "Avenir Next Condensed";
-  src: url("../fonts/Avenir Next Condensed.ttf.ttf");
-}
-
 .App {
   font-family: "Avenir Next Condensed";
   text-align: center;
